New method to handle the request; now i just pass the fd handler of the socket, and write over it.
This way its easier to create the handlers, and more efficient. On the bad side I trade off the
select solution, now olny one conection or threaded.

Added static handler that should work ok, and beginings of response objects. These objects are
just a helper; you can do that by hand if you want, but with the helper is easier to add headers
and just write the response header with onion_response_write. Actually almost all data is
properly filled.