Added onion_handler_opack, easy use of opack results. Fixed bug when sending data and break in the middle of send. oterm now uses internal resources instead of external, for !__DEBUG__. Added README.rst to oterm.
onion_handler_opack can be used to easy use opack results, only needs the opack writer and a path. Actually it can be used with your own writers, check onion_opack_render signature and do a funtion like that.

There was a bug when sending on SSL a big file, as it send max 16kB. Now it tries harder.

Opack now when compiled in non-debug mode returns the internal resources, that is, internal files, instead of using the directory handler.

The oterm README.rst file gives insight on how to use oterm, and where to find more information.