Refactored SSL. Now in stock library, but can be disabled with a flag, or when gnutls is not available.
The new SSL mode is just use the onion class, and if you need SSL set the certificates. You can check if
SSL support is available and enabled with the new onion_flags(), checking for O_SSL_AVAILABLE and O_SSL_ENABLED.

oterm uses the new way of using the SSL support, and added the flag --no-ssl.