Clarified the connection status return codes. Now all in the enum onion_connection_status.
All functions that need to set somehow the connection status (because of an error, because they know if it hsould wait for more data or keep alive) use this enum.

For values <0 it means close connection, for >=0 keep it open and feed more data to the request as data appears.