Onion requests now can have very big post data, set at max_post_data at onion_server. Smarter decoding and memory handling on requests.
Now on request all decoding is made in place, resulting in les smemory allocations and more locality. It copies the original data once (fast strdup), and decodes all in place.

For post if has a maximum full post size that is the coded size. If more it fails faster. If less it reads all the post, and decodes it in place too. All data freed correctly at end.