Fixed bug at POST files where it was writing \r\n on single \n found.
It was not beacuse of dos2unix stuff, but that the boundary starts
always \r\n-- but give the option to start on \n. If so it must
remember it did not start on \r in case its not the boundary which
is found.

Also related if start to find the boundary it may fail, but with a
real boundary. If ignoring this case on singe \n it just dont end,
and on \r\n it wrote \r at the end of file.