Added refcount support to sessions.
With this sessions should be fully functional: Each request that needs the session calls it and gets a reference to the dictionary.
On request deletion such reference is released and if all references are released the dict is really freed.
Each dict has a rwlock so on multithreaded environments it should rwlock as necesary when addig or accessing data.