Big rework of handlers: Now prefered and easy to use url_handler is an internal handler.
onion_url is the rename and extension of onion_handler_url, and now is the prefered way to set up the
server layout. It breaks a little the "onion" layer concept, but its necesary to make it easy to use.

The onion object can create it automatically and reuse if necesary, or can be used independently. It has
several helpers to add the most common cases: simple handlers, handlers with data, externally created handlers,
and static data.

It adds more complexity to the core, but it was shown that the other aproach is dificult and error prone. Anyway
this change do not discard the use of the other approach at all, and can still be used.

As now th path and regexp are no longer necesary they are removed and all examples and tests were changed.