Added helper onion_png_response that is able to return a generated image.
To test this functionality, the mandelbrot example ws added. This example shows a mandelbrot set on the browser,
with data calculated on the server. It is also made in C++, to prove interoperatibility. This
example was motivated by swill, a competing library.

Some more fixes around, as in otemplate, if no dictionary is passed, none is freed.