Added longpoll example and supporting functions.
An OCS_YIELD handler return code has been added, which forces
not to release the request nor response. This allows the poller
to regain the thread and keep processing, while with some extra work
keep sending data in another thread.

It required some extra functions on onion_ptr_list which is a linked
list of pointers to allow removing and filtering.

The poller has a new function to allow get a poller slot, required to
remove the free function.

Added fake onion_poller_get to libev/libevent pollers

It will fail miserably it anyone tries to use it.