Fix undefined behavior in dict.c.
The ctype macros isspace and isdigit take an int instead of a char, and
they expect the input to be representable as an unsigned char or the EOF
macro. In this case we were passing the char directly, which could lead
to undefined behavior. This issue was pointed out by Solar Designer on
the mailing list.