Fix timeout spurious segfault
This situation oculd happen before:

	1. Thread A: Timeout marked, stop before checking who
	2. Thread B: Data received on slot, stop before processing
	3. Thread A: Remove slot
	4. Thread B: Use free'd slot data

Now, step 3 does not remove the slot, but closes the connection, which
makes step 4 fail at write, or out of this situation another epoll_wait
to catch the closing of the fd. Data is only freed when detecting close
(EPOLLRDHUP remote HUP, or EPOLLHUP local HUP).

Thanks to Rich Felker and Alexander for the help on detecting, hunting,
and solving this bug.