Slots now use a id system for locating through epoll
There were problems of using epoll data as is or fd as both
can imply some data reuse after free, for example:

1. Thread A: Get epoll signal, get fd or data ptr
2. Thread B: Timeout, mark for deletion
3. Thread C: Delete
4. Thread D: Reuse fd
5. Thread A: Add data is expired, no longer fd valid, nor data ptr.

Using a secuential id as ev->data.u64, and a resizable array, its possible
to look always by this id, and if its not available, just dont return the
el. Also by marking the id as in use, no two threads will try to use the
slot at the same time.

Finally, removing by id is necesary as the fd can be reused, so searching
for the slot by fd could return an old or in process slot.