Allow external header management.
This introduces an option to have the framework hand over header management to
the application.  The header section will not be finished automatically and the
app can keep adding headers via the usual write functions. It also becomes the
app's responsibility to end the header section before sendind the body.

This option is useful to proxy requests to CGI apps, but can probably come in
handy in other situations as well.