Use proper init/deinit for static data at poller.
Use the poller_slot pool, but instead of ficed static always,
it is init at onion_poller_new and deinit at onion_poller_free.

It has a refcount to check when to proper deinit the data. Normaly
it is only used by one poller, but with the refcount we cover our
backs in case of several pollers. Refcount uses GCC atomics.