WIP #189. Initial API specification for hooks
I added some hooks into types.h and some initial API at onion.h and API
example use at 22-hooks.c.

This is just a start to decide the final API before implementing
anything.

My idea is that hooks can be added as required at these points, with any
number of hooks at any and the processed in insertion order.

The hook itself receives some user private data and both the request
and the response, and modify them as required. Essentially are handlers
and so this typedef is used.

It can return OCS_OK signalling to go on, or any other OCS_ value
(errors and so on).

Soem internal functions can be changed to use the hooks, as loggin
and error handling.

Please comment on the API and if it looks ok, I will continue with the
implementation.