Proposal for Issue 263 (basename() on Android)
This commit adds the code of GLibC's xpg_basename.c (LGPL)
and replaces basename() with __xpg_basename() in onions logging
function.
The basename() function of Androids Bionic Libc returns
a pointer to a static array. This is problematic if return variables
of two basename()-calls interleaves,
i.e. in ofileserver with DEBUG=True